+++ lib/msal-browser/src/controllers/StandardController.ts
@@ -28,6 +28,7 @@ import {
     AccountEntityUtils,
     Constants,
     AuthToken,
+    enforceResourceParameter,
 } from "@azure/msal-common/browser";
 import * as BrowserPerformanceEvents from "../telemetry/BrowserPerformanceEvents.js";
 import * as BrowserRootPerformanceEvents from "../telemetry/BrowserRootPerformanceEvents.js";
@@ -101,7 +102,7 @@ function preflightCheck(
 ) {
     try {
         BrowserUtils.preflightCheck(initialized);
-        BrowserUtils.enforceResourceParameter(config.auth.isMcp, request);
+        enforceResourceParameter(config.auth.isMcp, request);
     } catch (e) {
         performanceEvent.end({ success: false }, e, request.account);
         throw e;
@@ -647,10 +648,7 @@ export class StandardController implements IController {
 
         try {
             BrowserUtils.redirectPreflightCheck(this.initialized, this.config);
-            BrowserUtils.enforceResourceParameter(
-                this.config.auth.isMcp,
-                request
-            );
+            enforceResourceParameter(this.config.auth.isMcp, request);
             this.browserStorage.setInteractionInProgress(
                 true,
                 INTERACTION_TYPE.SIGNIN
