+++ lib/msal-node/docs/configuration.md
@@ -75,7 +75,7 @@ const msalInstance = new PublicClientApplication(msalConfig);
 | `authorityMetadata`      | A string containing the .well-known/openid-configuration endpoint response. See [Performance](../../msal-common/docs/performance.md) for more info                                                             | string                                                                                                                                       | Empty string `""`                                                          |
 | `clientCapabilities`     | Array of capabilities to be added to all network requests as part of the `xms_cc` claims request (see: [Client capability in MSAL](../../msal-common/docs/client-capability.md))                               | Array of strings                                                                                                                             | []                                                                         |
 | `azureCloudOptions`      | A defined set of azure cloud options for developers to default to their specific cloud authorities, for specific clouds supported please refer to the [AzureCloudInstance](aka.ms/msaljs/azure_cloud_instance) | [AzureCloudOptions](https://azuread.github.io/microsoft-authentication-library-for-js/ref/modules/_azure_msal_common.html#azurecloudoptions) | [AzureCloudInstance.None](msaljs/azure_cloud_instance)                     |
-| `isMcp`                  | If true, a `resource` parameter is required on all token requests. Used for MCP flows.                                                                                               | boolean                                                                                                                                      | `false`                                                                    |
+| `isMcp`                  | If true, a `resource` parameter is required on all token requests. Used for MCP flows. See [MCP documentation](mcp.md) for more details.                                                                                               | boolean                                                                                                                                      | `false`                                                                    |
 
 ### Cache Config Options
